fix: remove redundant .astype(np.float32) before .tolist()
Gemini thread PRRT_kwDORVL_j85-6L4w:
numpy's .tolist() converts float16 array elements directly to Python float
(64-bit doubles), making the intermediate .astype(np.float32) step
redundant. Removing it is numerically identical and avoids an unnecessary
array copy during every WAL decode.

+++ graphiti_core/driver/wal_replay_helpers.py
@@ -164,7 +164,7 @@ def decode_embedding_param(value: Any) -> Any:
         if value.startswith('f16:'):
             try:
                 raw = base64.b64decode(value[4:])
-                return np.frombuffer(raw, dtype=_F16).astype(np.float32).tolist()
+                return np.frombuffer(raw, dtype=_F16).tolist()
             except ValueError as e:
                 raise ValueError(f'Failed to decode f16: embedding payload: {e}') from e
         if value.startswith('f32:'):
